+++ raiden-ts/src/global.d.ts
@@ -13,5 +13,6 @@ declare module 'abort-controller/polyfill';
 declare module 'wrtc'; // types provided by @types/webrtc
 
 declare module 'pouchdb-adapter-indexeddb';
+declare module 'pouchdb-debug';
 
 declare type Mutable<T> = { -readonly [P in keyof T]: T[P] };